changed how the preogram work

+++ 12-binary_tree_leaves.c
@@ -9,7 +9,7 @@
 
 size_t binary_tree_leaves(const binary_tree_t *tree)
 {
-	int nleaves = 0;
+	size_t nleaves = 0;
 
 	if (tree == NULL)
 		return (0);
